Created authentication smoke test
This also fixes a side effect issue in tests where between ending a test
and starting another. The prior server was still running side effects which
caused the next server to fail (i.e. the search index was deleted in the
next server while search jobs were still being run in the prior).