Fixed bug causing Review type to be overwritten
This fixes an issue where the type of community item was being overwritten
elsewhere (in the photo gallery), causing the community item type to always be
a general community item, rather than a review for example. The type resolver
in photo gallery was always return general, regardless of what the community
item interface type resolver said it was. Relay was trusting the photo gallery
type resolver, thus overwriting the type.