Attempt at duplicate topic disambiguator
I tried adding a disambiguator field to the topics model, but was once again
stopped by bugs in Loopback. Loopback was not letting the disambiguator
column to be non-null with a default value of ''. This is the only way to
enforce a unique constraint on topic name+disambiguator since PostgreSQL does
not consider duplicate null values as unique. The only solution seems to take
Loopback to a field and beat the shit out of it with a baseball bat while
gangsta rap plays in the background.